requirements.txt
requests==2.31.0
python-dotenv==1.0.0
slack-sdk==3.23.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
discord.py==2.3.2